Move main logic to separate file
