+++ src/main.ts
@@ -1,0 +1,25 @@
+import glob from 'fast-glob';
+
+import lintFiles from './lint-files';
+
+const ERROR_CODE = 1;
+const SUCCESS_CODE = 0;
+const ignore = ['node_modules', 'README.md', 'CHANGELOG.md', 'LICENSE'];
+
+export default function main(): Promise<number> {
+  return glob('**/*', { ignore })
+    .then(files => {
+      const errorMessages = lintFiles(files);
+
+      if (errorMessages.length > 0) {
+        errorMessages.forEach(message => console.error(message));
+        return Promise.resolve(ERROR_CODE);
+      } else {
+        return Promise.resolve(SUCCESS_CODE);
+      }
+    })
+    .catch(error => {
+      console.error(error);
+      return Promise.resolve(ERROR_CODE);
+    });
+}
